+++ R/SoilProfileCollection-slice-methods.R
@@ -124,8 +124,8 @@ slice.fast <- function(object, fm, top.down=TRUE, just.the.data=FALSE, strict=TR
 
   # extract components of the formula:
   formula <- paste(deparse(fm, 500), collapse = "")
-  elements <- stri_split_fixed(str = formula, pattern = '~')[[1]]
-  formula <- lapply(stri_split_regex(str = elements, pattern = '[+*]'), trimws)
+  elements <- strsplit(x = fm, split = '~', fixed = )[[1]]
+  formula <- lapply(strsplit(x = elements, split = '[+*]'), trimws)
 
   # TODO: this will have to be changed when we implement no LHS = all slices
   if (length(formula) > 2)